move back server extracted networking functions back into common
