+++ common/src/y-network/utils.ts
@@ -1,0 +1,360 @@
+import * as Y from "yjs";
+
+import * as syncProtocol from "y-protocols/sync";
+import * as awarenessProtocol from "y-protocols/awareness";
+
+import { encoding, decoding, mutex, map } from "lib0";
+
+import debounce from "lodash.debounce";
+
+import { LeveldbPersistence } from "y-leveldb";
+
+import { defaultCallbackHandler, isDefaultCallbackSet } from "./callback.js";
+
+const CALLBACK_DEBOUNCE_WAIT =
+  parseInt(process.env.CALLBACK_DEBOUNCE_WAIT) || 2000;
+const CALLBACK_DEBOUNCE_MAXWAIT =
+  parseInt(process.env.CALLBACK_DEBOUNCE_MAXWAIT) || 10000;
+
+const wsReadyStateConnecting = 0;
+const wsReadyStateOpen = 1;
+const wsReadyStateClosing = 2; // eslint-disable-line
+const wsReadyStateClosed = 3; // eslint-disable-line
+
+// disable gc when using snapshots!
+const gcEnabled = process.env.GC !== "false" && process.env.GC !== "0";
+const persistenceDir = process.env.YPERSISTENCE;
+/**
+ * @type {{bindState: function(string,WSSharedDoc):void|Promise<void>, writeState:function(string,WSSharedDoc):Promise<any>, provider: any}|null}
+ */
+let persistence = null;
+if (typeof persistenceDir === "string") {
+  console.info('Persisting documents to "' + persistenceDir + '"');
+  // @ts-ignore
+  const ldb = new LeveldbPersistence(persistenceDir);
+  persistence = {
+    provider: ldb,
+    bindState: async (docName, ydoc) => {
+      const persistedYdoc = await ldb.getYDoc(docName);
+      const newUpdates = Y.encodeStateAsUpdate(ydoc);
+      ldb.storeUpdate(docName, newUpdates);
+      Y.applyUpdate(ydoc, Y.encodeStateAsUpdate(persistedYdoc));
+      ydoc.on("update", (update) => {
+        ldb.storeUpdate(docName, update);
+      });
+    },
+    writeState: async (docName, ydoc) => {},
+  };
+} else if (globalThis["global"]) {  // If running in node.js, `global` will exist.
+  console.warn("Missing YPERSISTENCE environment variable. y-websocket data will NOT be saved.");
+}
+
+/**
+ * @param {{bindState: function(string,WSSharedDoc):void,
+ * writeState:function(string,WSSharedDoc):Promise<any>,provider:any}|null} persistence_
+ */
+export const setPersistence = (persistence_) => {
+  persistence = persistence_;
+};
+
+/**
+ * @return {null|{bindState: function(string,WSSharedDoc):void,
+ * writeState:function(string,WSSharedDoc):Promise<any>}|null} used persistence layer
+ */
+export const getPersistence = () => persistence;
+
+/**
+ * @type {Map<string,WSSharedDoc>}
+ */
+export const docs = new Map();
+
+const messageSync = 0;
+const messageAwareness = 1;
+// const messageAuth = 2
+
+/**
+ * @param {Uint8Array} update
+ * @param {any} origin
+ * @param {WSSharedDoc} doc
+ */
+const updateHandler = (update, origin, doc) => {
+  const encoder = encoding.createEncoder();
+  encoding.writeVarUint(encoder, messageSync);
+  syncProtocol.writeUpdate(encoder, update);
+  const message = encoding.toUint8Array(encoder);
+  doc.conns.forEach((_, conn) => send(doc, conn, message));
+};
+
+class WSSharedDoc extends Y.Doc {
+  name: string;
+  mux: any;
+  conns: Map<Object, Set<number>>;
+  awareness: any;
+  whenSynced: Promise<void>;
+  isSynced: boolean;
+
+  /**
+   * @param {string} name
+   */
+  constructor(
+    name,
+    { callbackHandler = isDefaultCallbackSet ? defaultCallbackHandler : null }
+  ) {
+    super({ gc: gcEnabled });
+    this.name = name;
+    this.mux = mutex.createMutex();
+    /**
+     * Maps from conn to set of controlled user ids. Delete all user ids from awareness when this conn is closed
+     * @type {Map<Object, Set<number>>}
+     */
+    this.conns = new Map();
+    /**
+     * @type {awarenessProtocol.Awareness}
+     */
+    this.awareness = new awarenessProtocol.Awareness(this);
+    this.awareness.setLocalState(null);
+    /**
+     * @type {Promise<void>|void}
+     */
+    this.whenSynced = void 0;
+    /**
+     * @type {boolean}
+     */
+    this.isSynced = false;
+    /**
+     * @param {{ added: Array<number>, updated: Array<number>, removed: Array<number> }} changes
+     * @param {Object | null} conn Origin is the connection that made the change
+     */
+    const awarenessChangeHandler = ({ added, updated, removed }, conn) => {
+      const changedClients = added.concat(updated, removed);
+      if (conn !== null) {
+        const connControlledIDs =
+          /** @type {Set<number>} */ this.conns.get(conn);
+        if (connControlledIDs !== undefined) {
+          added.forEach((clientID) => {
+            connControlledIDs.add(clientID);
+          });
+          removed.forEach((clientID) => {
+            connControlledIDs.delete(clientID);
+          });
+        }
+      }
+      // broadcast awareness update
+      const encoder = encoding.createEncoder();
+      encoding.writeVarUint(encoder, messageAwareness);
+      encoding.writeVarUint8Array(
+        encoder,
+        awarenessProtocol.encodeAwarenessUpdate(this.awareness, changedClients)
+      );
+      const buff = encoding.toUint8Array(encoder);
+      this.conns.forEach((_, c) => {
+        send(this, c, buff);
+      });
+    };
+    this.awareness.on("update", awarenessChangeHandler);
+    this.on("update", updateHandler);
+    if (callbackHandler) {
+      this.on(
+        "update",
+        debounce(callbackHandler, CALLBACK_DEBOUNCE_WAIT, {
+          maxWait: CALLBACK_DEBOUNCE_MAXWAIT,
+        })
+      );
+    }
+    if (persistence !== null) {
+      this.whenSynced = persistence.bindState(name, this).then(() => {
+        this.isSynced = true;
+      });
+    }
+  }
+}
+
+/**
+ * Gets a Y.Doc by name, whether in memory or on disk
+ *
+ * @param {string} docname - the name of the Y.Doc to find or create
+ * @param {boolean} gc - whether to allow gc on the doc (applies only when created)
+ * @return {WSSharedDoc}
+ */
+export const getYDoc = (docname, { gc = true, callbackHandler = null } = {}) =>
+  map.setIfUndefined(docs, docname, () => {
+    const doc = new WSSharedDoc(docname, { callbackHandler });
+    doc.gc = gc;
+    docs.set(docname, doc);
+    return doc;
+  });
+
+/**
+ * @param {any} conn
+ * @param {WSSharedDoc} doc
+ * @param {Uint8Array} message
+ */
+const messageListener = async (conn, doc, message) => {
+  try {
+    const encoder = encoding.createEncoder();
+    const decoder = decoding.createDecoder(message);
+    const messageType = decoding.readVarUint(decoder);
+    switch (messageType) {
+      case messageSync:
+        // await the doc state being updated from persistence, if available, otherwise
+        // we may send sync step 2 too early
+        if (doc.whenSynced) {
+          await doc.whenSynced;
+        }
+        encoding.writeVarUint(encoder, messageSync);
+        syncProtocol.readSyncMessage(decoder, encoder, doc, null);
+        if (encoding.length(encoder) > 1) {
+          send(doc, conn, encoding.toUint8Array(encoder));
+        }
+        break;
+      case messageAwareness: {
+        awarenessProtocol.applyAwarenessUpdate(
+          doc.awareness,
+          decoding.readVarUint8Array(decoder),
+          conn
+        );
+        break;
+      }
+    }
+  } catch (err) {
+    console.error(err);
+    doc.emit("error", [err]);
+  }
+};
+
+/**
+ * @param {WSSharedDoc} doc
+ * @param {any} conn
+ */
+const closeConn = (doc, conn) => {
+  if (doc.conns.has(conn)) {
+    /**
+     * @type {Set<number>}
+     */
+    // @ts-ignore
+    const controlledIds = doc.conns.get(conn);
+    doc.conns.delete(conn);
+    awarenessProtocol.removeAwarenessStates(
+      doc.awareness,
+      Array.from(controlledIds),
+      null
+    );
+    if (doc.conns.size === 0 && persistence !== null) {
+      if (doc.isSynced) {
+        // if persisted and the state has finished loading from the database,
+        // we write the state back to persisted storage
+        persistence.writeState(doc.name, doc).then(() => {
+          doc.destroy();
+        });
+      }
+      docs.delete(doc.name);
+    }
+  }
+  conn.close();
+};
+
+/**
+ * @param {WSSharedDoc} doc
+ * @param {any} conn
+ * @param {Uint8Array} m
+ */
+const send = (doc, conn, m) => {
+  if (
+    conn.readyState !== wsReadyStateConnecting &&
+    conn.readyState !== wsReadyStateOpen
+  ) {
+    closeConn(doc, conn);
+  }
+  try {
+    conn.send(
+      m,
+      /** @param {any} err */ (err) => {
+        err != null && closeConn(doc, conn);
+      }
+    );
+  } catch (e) {
+    closeConn(doc, conn);
+  }
+};
+
+const pingTimeout = 30000;
+
+/**
+ * @param {any} conn
+ * @param {any} req
+ * @param {any} opts
+ */
+export const setupWSConnection = async (
+  conn,
+  req,
+  {
+    docName = req.url.slice(1).split("?")[0],
+    gc = true,
+    callbackHandler = null,
+  } = {}
+) => {
+  conn.binaryType = "arraybuffer";
+  // get doc, initialize if it does not exist yet
+  const doc = getYDoc(docName, { gc, callbackHandler });
+  doc.conns.set(conn, new Set());
+  // listen and reply to events
+  conn.on(
+    "message",
+    /** @param {ArrayBuffer} message */ (message) =>
+      messageListener(conn, doc, new Uint8Array(message))
+  );
+
+  // Check if connection is still alive
+  let pongReceived = true;
+  const pingInterval = setInterval(() => {
+    if (!pongReceived) {
+      if (doc.conns.has(conn)) {
+        closeConn(doc, conn);
+      }
+      clearInterval(pingInterval);
+    } else if (doc.conns.has(conn)) {
+      pongReceived = false;
+      try {
+        conn.ping();
+      } catch (e) {
+        closeConn(doc, conn);
+        clearInterval(pingInterval);
+      }
+    }
+  }, pingTimeout);
+  conn.on("close", () => {
+    closeConn(doc, conn);
+    clearInterval(pingInterval);
+  });
+  conn.on("pong", () => {
+    pongReceived = true;
+  });
+  // put the following in a variables in a block so the interval handlers don't keep in in
+  // scope
+  {
+    // await the doc state being updated from persistence, if available, otherwise
+    // we may send sync step 1 too early
+    if (doc.whenSynced) {
+      await doc.whenSynced;
+    }
+
+    // send sync step 1
+    const encoder = encoding.createEncoder();
+    encoding.writeVarUint(encoder, messageSync);
+    syncProtocol.writeSyncStep1(encoder, doc);
+    send(doc, conn, encoding.toUint8Array(encoder));
+    const awarenessStates = doc.awareness.getStates();
+    if (awarenessStates.size > 0) {
+      const encoder = encoding.createEncoder();
+      encoding.writeVarUint(encoder, messageAwareness);
+      encoding.writeVarUint8Array(
+        encoder,
+        awarenessProtocol.encodeAwarenessUpdate(
+          doc.awareness,
+          Array.from(awarenessStates.keys())
+        )
+      );
+      send(doc, conn, encoding.toUint8Array(encoder));
+    }
+  }
+};
