+++ server/src/routes/relm.ts
@@ -35,8 +35,8 @@ import {
   respondWithSuccess,
   respondWithError,
   randomToken,
-  getYDoc
 } from "../utils/index.js";
+import { getYDoc } from "../getYDoc.js";
 
 const conscript = mkConscript();
 
@@ -173,7 +173,7 @@ relm.post(
         }
       } else {
         // Populate the new relm with basic ground
-        const newRelmDoc: Y.Doc = await getYDoc(relm.permanentDocId, { storeInCache: false });
+        const newRelmDoc: Y.Doc = await getYDoc(relm.permanentDocId);
         importWorldDoc(config.DEFAULT_RELM_CONTENT, newRelmDoc);
 
         console.log(
@@ -191,10 +191,10 @@ relm.post(
 );
 
 async function cloneRelmDoc(seedRelmDocId: string, newRelmDocId: string) {
-  const seedRelmDoc: Y.Doc = await getYDoc(seedRelmDocId, { storeInCache: false });
+  const seedRelmDoc: Y.Doc = await getYDoc(seedRelmDocId);
   const relmContent = exportWorldDoc(seedRelmDoc);
 
-  const newRelmDoc: Y.Doc = await getYDoc(newRelmDocId, { storeInCache: false });
+  const newRelmDoc: Y.Doc = await getYDoc(newRelmDocId);
   importWorldDoc(relmContent, newRelmDoc);
 
   return newRelmDoc;
@@ -223,7 +223,7 @@ relm.post(
   middleware.authenticated(),
   middleware.authorized("access"),
   wrapAsync(async (req, res) => {
-    const relmDoc = await getYDoc(req.relm.permanentDocId, { storeInCache: false });
+    const relmDoc = await getYDoc(req.relm.permanentDocId);
     const content = exportWorldDoc(relmDoc);
 
     return respondWithSuccess(res, {
@@ -241,7 +241,7 @@ relm.post(
   middleware.authenticated(),
   middleware.authorized("access"),
   wrapAsync(async (req, res) => {
-    const doc: Y.Doc = await getYDoc(req.relm.permanentDocId, { storeInCache: false });
+    const doc: Y.Doc = await getYDoc(req.relm.permanentDocId);
     req.relm.permanentDocSize = Y.encodeStateAsUpdate(doc).byteLength;
     const twilioToken = twilio.getToken(req.authenticatedParticipantId);
 
@@ -326,7 +326,7 @@ relm.post(
   middleware.acceptJwt(),
   middleware.authorized("read"),
   wrapAsync(async (req, res) => {
-    const doc: Y.Doc = await getYDoc(req.relm.permanentDocId, { storeInCache: false });
+    const doc: Y.Doc = await getYDoc(req.relm.permanentDocId);
     req.relm.permanentDocSize = Y.encodeStateAsUpdate(doc).byteLength;
 
     // TODO: don't reveal twilio token for `read`-level permission
@@ -407,7 +407,7 @@ relm.post(
       return respondWithError(res, "changes required");
     }
 
-    const doc: Y.Doc = await getYDoc(req.relm.permanentDocId, { storeInCache: false });
+    const doc: Y.Doc = await getYDoc(req.relm.permanentDocId);
     // console.log("doc.actions", doc.getMap("actions").toJSON());
 
     const variables = await Variable.getVariables({ relmId });
@@ -492,7 +492,7 @@ relm.post(
   middleware.acceptJwt(),
   middleware.authorized("edit"),
   wrapAsync(async (req, res) => {
-    const doc: Y.Doc = await getYDoc(req.relm.permanentDocId, { storeInCache: false });
+    const doc: Y.Doc = await getYDoc(req.relm.permanentDocId);
 
     const actions = req.body.actions;
